Fix data source slug update (#204)
When updating the new sanitized slug in form state it was compared with
the unsanitized version of it in the local state of SlugInput component.
This led to cases when the sanitized and unsanitized version being same,
the changed slug was not updated in form state leading to it not being
saved when 'Save' button on form was clicked. Comparing the local
sanitized version with slug in the form state fixes it.